refactor(net): split poll_fd into cfg-inners
